# backend/swagger.json
{"swagger":"2.0","info":{"description":"hangTogether 서버 api","version":"1.0.0","title":"hangTogether API","contact":{"email":"lmjing1234@gmail.com"},"termsOfService":"http://ec2-13-124-183-2.ap-northeast-2.compute.amazonaws.com:3000/","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"tags":[{"name":"user","description":"foreigner & korean 유저 정보"},{"name":"language","description":"language list"},{"name":"post","description":"모집 글"}],"paths":{"/user/check":{"get":{"tags":["user"],"summary":"이메일 & 닉네임 중복확인","operationId":"checkDuplication","description":"이메일 or 닉네임 중 중복확인 하고자 하는 것을 query에 담아서 보내주면 됨. 동시에 확인 불가!  예시 - /user/check?email=lmjing1234@gmail.com","produces":["application/json"],"parameters":[{"in":"query","name":"email","description":"이메일 중복확인 시 필수","required":false,"type":"string"},{"in":"query","name":"nickname","description":"닉네임 중복확인 시 필수","required":false,"type":"string"}],"responses":{"200":{"description":"사용 가능"},"400":{"description":"bad input parameter"},"409":{"description":"an existing item already exists"},"500":{"description":"internal server error"}}}},"/user":{"post":{"tags":["user"],"summary":"회원가입","operationId":"joinUser","description":"회원가입 실패 시 메세지에 내용 들어있으니 그걸 toast나 snackbar 메세지에 띄워놓을것","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"joinUser","description":"언어는 한개 이상 3개 이하 / profileUrl과 introduce는 설정하지 않았을 경우 보내지 않아도 됨","required":true,"schema":{"$ref":"#/definitions/JoinUser"}}],"responses":{"201":{"description":"성공적으로 생성됨","schema":{"$ref":"#/definitions/User"}},"400":{"description":"invalid input, object invalid."},"409":{"description":"an existing item already exists"},"500":{"description":"internal server error"}}}},"/user/{id}":{"get":{"tags":["user"],"summary":"특정 유저 정보 조회","operationId":"getUserInfo","description":"유저 아이디로 특정 유저 정보 조회함. 외국인, 한국인 타입에 따라 받는 정보가 다를 수 있음","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"유저의 _id","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/User"}},"400":{"description":"bad input parameter"},"404":{"description":"not found"},"500":{"description":"internal server error"}}},"put":{"tags":["user"],"summary":"회원 정보 수정","operationId":"editUser","description":"닉네임, 자기소개, 비밀 번호 변경","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"joinUser","description":"변경하지 않은 정보(기존과 같은 정보)가 있어도 넣어서 보내줘야 함","required":true,"schema":{"$ref":"#/definitions/EditUser"}},{"in":"path","name":"id","description":"유저의 _id","required":true,"type":"string"}],"responses":{"201":{"description":"성공적으로 수정 됨","schema":{"$ref":"#/definitions/User"}},"400":{"description":"invalid input, object invalid."}}},"delete":{"tags":["user"],"summary":"유저 탈퇴","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"유저의 _id","required":true,"type":"string"}],"responses":{"204":{"description":"Success - Not content"},"400":{"description":"Bad request. invalid parameter or invalid body"},"404":{"description":"Not found. 그런 유저 없습니다."}}}},"/language":{"get":{"tags":["language"],"summary":"언어 목록 조회","operationId":"getLanguageList","description":"유저 사용 가능 언어 목록","produces":["application/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/Language"}}},"404":{"description":"not found"}}}},"/post":{"get":{"tags":["post"],"summary":"모집중인 게시글 목록 조회","operationId":"getPostList","description":"옵션 적용한 경우 해당 옵션 쿼리문에 날려 보내기(옵션이 무관인 경우 null이나 입력하지 않고 보내기부) / 그냥 메인에서 전체 목록 부를 경우엔 쿼리문에 값 입력할 필요 X / 모집 중인 게시글만 확인 가능! / 한꺼번에 너무 많은 양을 부르면 문제가 될 수 있으니 페이지 처리 할거야 ( 페이스북 스크롤하면 더 불러오는 것 처럼 구현해야 함) 처음엔 10개만 그다음에 로드모어 요청 했을 경우 10개 더","produces":["application/json"],"parameters":[{"in":"query","name":"page","description":"맨 처음 페이지 = 0 / 그 다음 유저가 스크롤 한번해서 로드 모어 될 경우 +1씩 요청하길","required":true,"type":"integer"},{"in":"query","name":"age","description":"10 : 10대 / 20 : 20대 / 30 : 30대 / 40 : 40대 이상 ","required":false,"type":"integer"},{"in":"query","name":"startDate","description":"여행 시작일 / 입력시 endDate 입력 필수","required":false,"type":"string","format":"date"},{"in":"query","name":"endDate","description":"여행 마지막일 / 입력시 startDate 입력 필수","required":false,"type":"string","format":"date"},{"in":"query","name":"sex","description":"성별 male or female","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/Post"}}},"404":{"description":"not found"}}},"post":{"tags":["post"],"summary":"모집글 생성","operationId":"createPost","description":"외국인이 등록하는 모집 글","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"newPost","description":"content 내용은 필수 x / 여행지는 1개이상","required":true,"schema":{"$ref":"#/definitions/NewPost"}}],"responses":{"201":{"description":"성공적으로 생성됨","schema":{"$ref":"#/definitions/Post"}},"400":{"description":"invalid input, object invalid."},"409":{"description":"an existing item already exists"}}}},"/post/{id}":{"get":{"tags":["post"],"summary":"게시물 정보 조회","operationId":"getPost","description":"특정 게시글 조회","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"게시글 _id","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Post"}},"400":{"description":"bad input parameter"}}},"put":{"tags":["post"],"summary":"게시글 정보 수정","operationId":"editPost","description":"하고싶은 말, 여행 일정, 여행 날짜","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"게시글 _id","required":true,"type":"string"},{"in":"body","name":"newPost","description":"content 내용은 필수 x / 여행지는 1개이상","required":true,"schema":{"$ref":"#/definitions/EditPost"}}],"responses":{"201":{"description":"성공적으로 수정 됨","schema":{"$ref":"#/definitions/Post"}},"400":{"description":"invalid input, object invalid."}}},"delete":{"tags":["post"],"summary":"게시글 마감","description":"지원자 선택 여부와 상관없이 마감됨","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"게시글 _id","required":true,"type":"string"}],"responses":{"201":{"description":"Success - 마감됨"},"400":{"description":"Bad request. invalid parameter or invalid body"},"404":{"description":"Not found. 게시물을 찾을 수 없습니다"}}}},"/post/{id}/apply":{"put":{"tags":["post"],"summary":"가이드 지원","operationId":"applyPost","description":"하고싶은 말, 여행 일정, 여행 날짜(서버에서 따로 예외처리 안하니깐 업로드한 날짜에 해당되는지 확인 할 것!)","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"게시글 _id","required":true,"type":"string"},{"in":"body","name":"volunteer","description":"여행 일정은 적어도 1개 / 메시지는 null 가능","required":true,"schema":{"$ref":"#/definitions/Volunteer"}}],"responses":{"201":{"description":"성공적으로 지원 됨"},"400":{"description":"invalid input, object invalid."}}},"delete":{"tags":["post"],"summary":"지원 취소","operationId":"deleteApplyPost","description":"쿼리 user에 유저 아이디 반드시 보내야 함","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"게시글 _id","required":true,"type":"string"},{"in":"query","name":"user","description":"user의 _id","required":true,"type":"string"}],"responses":{"201":{"description":"Success - 신청이 취소되었습니다."},"409":{"description":"Fail - 신청에 실패했습니다."},"500":{"description":"internal server error"}}}},"/post/user/{id}":{"get":{"tags":["post"],"summary":"유저의 게시글 목록(히스토리) 조회","operationId":"getUserPostList","description":"마감된 정보도 확인 가능 / 최신 순","produces":["application/json"],"parameters":[{"in":"path","name":"id","description":"user의 _id","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/Post"}}},"404":{"description":"not found"}}}}},"definitions":{"JoinUser":{"type":"object","required":["email","password","nickname","sex","birth","type","languages"],"properties":{"email":{"type":"string","format":"email","example":"lmjing1234@gmail.com"},"password":{"type":"string","format":"길이 제한이나 영어 숫자만 가능같은거 할까?","example":"abcde12345!"},"nickname":{"type":"string","example":"lmjing"},"sex":{"type":"string","example":"male or female","default":"male"},"birth":{"type":"string","format":"date","example":"1994-05-29"},"profileUrl":{"type":"string","format":"directory + png, jpg","example":"/resource/lmjing.png"},"type":{"type":"string","example":"korean or foreigner","default":"korean"},"languages":{"type":"array","description":"최소 1개 최대 3개","items":{"type":"string","example":"Korean"}},"introduce":{"type":"string","example":"hello! nice to meet you :)"}}},"User":{"type":"object","required":["email","password","nickname","sex","birth","type","languages"],"properties":{"_id":{"type":"string","example":"59cfc8f360d644190d2d43da","description":"DB에서 자동생성됨 이게 조회하기 쉬워서 이걸로 하니 이해 바람"},"email":{"type":"string","format":"email","example":"lmjing1234@gmail.com"},"password":{"type":"string","format":"길이 제한이나 영어 숫자만 가능같은거 할까?","example":"abcde12345!"},"nickname":{"type":"string","example":"lmjing"},"sex":{"type":"string","example":"male or female","default":"male"},"birth":{"type":"string","format":"date","example":"1994-05-29"},"profileUrl":{"type":"string","format":"directory + png, jpg","example":"/resource/lmjing.png"},"type":{"type":"string","example":"korean or foreigner","default":"korean"},"languages":{"type":"array","description":"최소 1개 최대 3개","items":{"$ref":"#/definitions/Language"}},"introduce":{"type":"string","example":"hello! nice to meet you :)"}}},"EditUser":{"type":"object","required":["password","nickname","introduce"],"properties":{"password":{"type":"string","format":"길이 제한이나 영어 숫자만 가능같은거 할까?","example":"abcde12345!"},"nickname":{"type":"string","example":"lmjing"},"introduce":{"type":"string","example":"hello! nice to meet you :)"}}},"Language":{"type":"object","properties":{"_id":{"type":"string","example":"59cfdd1943ac6022c14ed36f","description":"DB에서 자동생성됨"},"language":{"type":"string","example":"korean"}}},"Post":{"type":"object","required":["writer","tripDate","trip"],"properties":{"_id":{"type":"string","example":"59cfc8f360d644190d2d43da","description":"DB에서 자동생성됨"},"writer":{"$ref":"#/definitions/User"},"created":{"type":"string","format":"datetime","description":"작성 날짜","example":"2016-05-18T16:00:00Z","default":"Date.now"},"content":{"type":"string","default":null,"description":"소개, 하고싶은 말","example":"역사에 대해 자세하게 알려주실 분만 구합니다."},"tripDate":{"type":"object","description":"여행 일정(시작~끝)","required":["start","end"],"properties":{"start":{"type":"string","format":"date","example":"2016-05-18"},"end":{"type":"string","format":"date","example":"2016-05-18"}}},"trip":{"type":"array","description":"최소 1개","items":{"type":"object","description":"장소는 필수","required":["place"],"properties":{"date":{"type":"string","format":"date","example":"2018-12-30"},"place":{"type":"object","required":["name","address"],"properties":{"name":{"type":"string","example":"경복궁"},"place":{"type":"string","example":"서울시 00구 00동 12-21"}}}}}},"recuriting":{"type":"boolean","default":true,"description":"모집 여부(현재 상태)"},"guide":{"type":"array","description":"확정된 가이드","default":[],"items":{"type":"object","$ref":"#/definitions/User"}},"volunteer":{"type":"object","required":["user","dates"],"properties":{"user":{"type":"object","$ref":"#/definitions/User"},"message":{"type":"string","default":null,"example":"안녕하세요."},"dates":{"type":"array","description":"적어도 1개 이상","items":{"type":"string","format":"date","example":"2017-12-30"}}}}}},"NewPost":{"type":"object","required":["tripDate","trip"],"properties":{"writer":{"type":"string","description":"user _id","example":"59cfdd1943ac6022c14ed36f"},"content":{"type":"string","default":null,"description":"소개, 하고싶은 말","example":"역사에 대해 자세하게 알려주실 분만 구합니다."},"tripDate":{"type":"object","description":"여행 일정(시작~끝)","required":["start","end"],"properties":{"start":{"type":"string","format":"date","example":"2016-05-18"},"end":{"type":"string","format":"date","example":"2016-05-18"}}},"trip":{"type":"array","description":"최소 1개","items":{"type":"object","description":"장소는 필수","required":["place"],"properties":{"date":{"type":"string","format":"date"},"place":{"type":"object","required":["name","address"],"properties":{"name":{"type":"string","example":"경복궁"},"place":{"type":"string","example":"서울시 00구 00동 12-21"}}}}}}}},"EditPost":{"type":"object","required":["tripDate","trip"],"properties":{"content":{"type":"string","default":null,"description":"소개, 하고싶은 말","example":"역사에 대해 자세하게 알려주실 분만 구합니다."},"tripDate":{"type":"object","description":"여행 일정(시작~끝)","required":["start","end"],"properties":{"start":{"type":"string","format":"date","example":"2016-05-18"},"end":{"type":"string","format":"date","example":"2016-05-18"}}},"trip":{"type":"array","description":"최소 1개","items":{"type":"object","description":"장소는 필수","required":["place"],"properties":{"date":{"type":"string","format":"date"},"place":{"type":"object","required":["name","address"],"properties":{"name":{"type":"string","example":"경복궁"},"place":{"type":"string","example":"서울시 00구 00동 12-21"}}}}}}}},"History":{"type":"object","properties":{"ing":{"type":"array","description":"모집중인 내역","items":{"type":"object","$ref":"#/definitions/Post"}},"end":{"type":"array","description":"마감된 내역","items":{"type":"object","$ref":"#/definitions/Post"}}}},"Volunteer":{"type":"object","required":["volunteer"],"properties":{"volunteer":{"type":"string","description":"user _id","example":"59cfdd1943ac6022c14ed36f"},"message":{"type":"string","example":"가이드 지원합니다."},"dates":{"type":"array","description":"적어도 1개","items":{"type":"string","format":"date","example":"2018-03-28"}}}}},"schemes":["http"],"host":"ec2-13-124-183-2.ap-northeast-2.compute.amazonaws.com:3000"}